pyproject.toml
[project]
name = "workflow-tools"
version = "0.1.0"
description = "FastAPI backend for AI-powered workflow automation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.116.1,<1.0.0",
  "openai>=2.6.1,<3.0.0",
  "pydantic>=2.11.7,<3.0.0",
  "pyyaml>=6.0.2,<7.0.0",
  "uvicorn>=0.35.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.1,<1.0.0",
  "pytest>=8.4.2,<9.0.0",
]

[build-system]
requires = ["setuptools>=80.9.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
